requirements.txt
apache-airflow-providers-postgres
psycopg2-binary
huggingface-hub
